requirements.txt
numpy==1.26.3
pandas
pytorch
